Refactor code to use Javascript classes
