+++ index.html
@@ -20,7 +20,7 @@
     </div>
     <p id="error"></p>
 
-    <script src="./js/main.js"></script>
+    <script type="module" src="./js/main.js"></script>
 </body>
 
 </html>
